- Migrating to Portable Class Libraries   - WSCT.Wrapper.Core PCL project added (bridge interfaces with PCL)   - WSCT.Wrapper for Windows, Linux x32, Linux x64 and MacOSX split in independant native projects   - WSCT.Helpers   - WSCT.Stacks   - WSCT.ISO7816

+++ WSCT.ISO7816/AnswerToReset/ATR.cs
@@ -197,7 +197,7 @@ namespace WSCT.ISO7816.AnswerToReset
 
             // Historic bytes
             HistoricBytes = new byte[t0.Value & 0x0F];
-            Array.Copy(atr, offset, HistoricBytes, 0, HistoricBytes.Length);
+            Array.Copy(atr, (int)offset, HistoricBytes, 0, HistoricBytes.Length);
             offset += (uint)HistoricBytes.Length;
 
             // TCK
